requirements.txt
pandas>=2.0.3
numpy>=1.24.3
yfinance>=0.2.18
SQLAlchemy>=2.0.19
pyyaml>=6.0
statsmodels>=0.14.0
pytest>=8.0
matplotlib>=3.7.2
flake8>=7.1.1
fredapi
black
autoflake
isort
python-dotenv
pytest-mock
quantstats
vectorbt
alpha-vantage
